SME-18 move category list fetching in content detail and content list containers
