+++ pwa-shell/src/ui/content-detail/ContentDetailContainer.js
@@ -1,3 +1,4 @@
+import { get } from 'lodash';
 import { connect } from 'react-redux';
 import {
   getSelectedContent,
@@ -7,7 +8,11 @@ import {
   getPreviousFromSelectedContent,
 } from 'state/content/selectors';
 import { setRequiresAuth } from 'state/content/actions';
-import { fetchContentDetail, fetchProtectedContentDetail } from 'state/thunks';
+import {
+  fetchContentDetail,
+  fetchProtectedContentDetail,
+  fetchCategoryListAndFilters,
+} from 'state/thunks';
 import ContentDetail from 'ui/content-detail/ContentDetail';
 import { getSelectedContentType } from 'state/contentType/selectors';
 import { setSelectedContentType } from 'state/contentType/actions';
@@ -21,7 +26,7 @@ export const mapStateToProps = state => ({
   prevContent: getPreviousFromSelectedContent(state),
 });
 
-export const mapDispatchToProps = dispatch => ({
+export const mapDispatchToProps = (dispatch, ownProps) => ({
   fetchContentDetail: (location, params) => {
     const requiresAuth =
       new URLSearchParams(location.search).get('requiresAuth') === 'true';
@@ -33,6 +38,11 @@ export const mapDispatchToProps = dispatch => ({
     } else {
       dispatch(fetchContentDetail(id));
     }
+  },
+  fetchCategoryListAndFilters: () => {
+    const contentType = get(ownProps, 'match.params.contentType');
+    dispatch(setSelectedContentType(contentType));
+    dispatch(fetchCategoryListAndFilters());
   },
 });
